# Attached-Assets/Attached-Assets/artifacts/api-server/src/routes/complaints.ts
import { Router, type IRouter } from "express";
import { eq, desc, sql, and } from "drizzle-orm";
import { db, complaintsTable, timelineEventsTable } from "@workspace/db";
import {
  ListComplaintsQueryParams,
  CreateComplaintBody,
  UpdateComplaintBody,
  GetRecentComplaintsQueryParams,
} from "@workspace/api-zod";

const router: IRouter = Router();

const DEPARTMENT_MAP: Record<string, string> = {
  "Pothole": "Public Works Department",
  "Garbage": "Sanitation Department",
  "Water Leakage": "Water Supply Department",
  "Broken Streetlight": "Electricity Department",
  "Illegal Construction": "Town Planning Department",
  "Sewage Overflow": "Sewage & Drainage Department",
};

const SEVERITY_MAP: Record<string, string> = {
  "Pothole": "High",
  "Garbage": "Medium",
  "Water Leakage": "High",
  "Broken Streetlight": "Medium",
  "Illegal Construction": "Critical",
  "Sewage Overflow": "Critical",
};

function formatComplaint(c: typeof complaintsTable.$inferSelect) {
  return {
    ...c,
    createdAt: c.createdAt.toISOString(),
    updatedAt: c.updatedAt.toISOString(),
    resolvedAt: c.resolvedAt ? c.resolvedAt.toISOString() : null,
  };
}

router.get("/complaints/stats/summary", async (_req, res): Promise<void> => {
  const all = await db.select().from(complaintsTable);
  const pending = all.filter(c => c.status === "Pending").length;
  const inProgress = all.filter(c => c.status === "In Progress").length;
  const resolved = all.filter(c => c.status === "Resolved").length;
  const critical = all.filter(c => c.severity === "Critical").length;
  const today = new Date(); today.setHours(0, 0, 0, 0);
  const todayCount = all.filter(c => new Date(c.createdAt) >= today).length;
  const resolvedWithTime = all.filter(c => c.status === "Resolved" && c.resolvedAt);
  const avgDays = resolvedWithTime.length > 0
    ? resolvedWithTime.reduce((acc, c) => {
        const diff = (new Date(c.resolvedAt!).getTime() - new Date(c.createdAt).getTime()) / (1000 * 60 * 60 * 24);
        return acc + diff;
      }, 0) / resolvedWithTime.length
    : 4.2;
  res.json({
    total: all.length,
    pending,
    inProgress,
    resolved,
    avgResolutionDays: Math.round(avgDays * 10) / 10,
    criticalCount: critical,
    todayCount,
    satisfactionRate: 0.87,
  });
});

router.get("/complaints/stats/recent", async (req, res): Promise<void> => {
  const params = GetRecentComplaintsQueryParams.safeParse(req.query);
  const limit = params.success ? (params.data.limit ?? 10) : 10;
  const complaints = await db.select().from(complaintsTable)
    .orderBy(desc(complaintsTable.createdAt))
    .limit(limit);
  res.json(complaints.map(formatComplaint));
});

router.get("/complaints", async (req, res): Promise<void> => {
  const params = ListComplaintsQueryParams.safeParse(req.query);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const { status, department, severity, category, ward } = params.data;
  let query = db.select().from(complaintsTable).$dynamic();
  const conditions = [];
  if (status) conditions.push(eq(complaintsTable.status, status));
  if (department) conditions.push(eq(complaintsTable.department, department));
  if (severity) conditions.push(eq(complaintsTable.severity, severity));
  if (category) conditions.push(eq(complaintsTable.category, category));
  if (ward) conditions.push(eq(complaintsTable.ward, ward));
  if (conditions.length > 0) query = query.where(and(...conditions));
  const complaints = await query.orderBy(desc(complaintsTable.createdAt));
  res.json(complaints.map(formatComplaint));
});

router.post("/complaints", async (req, res): Promise<void> => {
  const parsed = CreateComplaintBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  const data = parsed.data;
  const department = DEPARTMENT_MAP[data.category] || "General Administration";
  const severity = SEVERITY_MAP[data.category] || "Medium";
  const confidence = 0.85 + Math.random() * 0.12;
  const aiSummary = `AI classified this as a ${data.category} issue in ${data.ward}. Routed to ${department} with ${severity.toLowerCase()} severity based on historical patterns.`;
  const [complaint] = await db.insert(complaintsTable).values({
    ...data,
    department,
    severity,
    aiConfidence: Math.round(confidence * 100) / 100,
    aiSummary,
    estimatedResolutionDays: severity === "Critical" ? 3 : severity === "High" ? 5 : 7,
  }).returning();
  await db.insert(timelineEventsTable).values({
    complaintId: complaint.id,
    eventType: "submitted",
    description: `Complaint submitted by ${data.citizenName}`,
    actor: data.citizenName,
  });
  await db.insert(timelineEventsTable).values({
    complaintId: complaint.id,
    eventType: "ai_routed",
    description: `AI routed to ${department} with ${Math.round(confidence * 100)}% confidence`,
    actor: "AI Engine",
  });
  res.status(201).json(formatComplaint(complaint));
});

router.get("/complaints/:id", async (req, res): Promise<void> => {
  const raw = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const id = parseInt(raw, 10);
  if (isNaN(id)) { res.status(400).json({ error: "Invalid ID" }); return; }
  const [complaint] = await db.select().from(complaintsTable).where(eq(complaintsTable.id, id));
  if (!complaint) { res.status(404).json({ error: "Not found" }); return; }
  res.json(formatComplaint(complaint));
});

router.patch("/complaints/:id", async (req, res): Promise<void> => {
  const raw = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const id = parseInt(raw, 10);
  if (isNaN(id)) { res.status(400).json({ error: "Invalid ID" }); return; }
  const parsed = UpdateComplaintBody.safeParse(req.body);
  if (!parsed.success) { res.status(400).json({ error: parsed.error.message }); return; }
  const updateData: Partial<typeof complaintsTable.$inferInsert> = { ...parsed.data };
  if (parsed.data.status === "Resolved") updateData.resolvedAt = new Date();
  const [complaint] = await db.update(complaintsTable).set(updateData).where(eq(complaintsTable.id, id)).returning();
  if (!complaint) { res.status(404).json({ error: "Not found" }); return; }
  if (parsed.data.status) {
    await db.insert(timelineEventsTable).values({
      complaintId: id,
      eventType: "status_changed",
      description: `Status updated to ${parsed.data.status}`,
      actor: "Government Officer",
    });
  }
  res.json(formatComplaint(complaint));
});

router.post("/complaints/:id/vote", async (req, res): Promise<void> => {
  const raw = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const id = parseInt(raw, 10);
  if (isNaN(id)) { res.status(400).json({ error: "Invalid ID" }); return; }
  const [complaint] = await db.update(complaintsTable)
    .set({ votes: sql`${complaintsTable.votes} + 1` })
    .where(eq(complaintsTable.id, id))
    .returning();
  if (!complaint) { res.status(404).json({ error: "Not found" }); return; }
  res.json(formatComplaint(complaint));
});

router.get("/complaints/:id/timeline", async (req, res): Promise<void> => {
  const raw = Array.isArray(req.params.id) ? req.params.id[0] : req.params.id;
  const id = parseInt(raw, 10);
  if (isNaN(id)) { res.status(400).json({ error: "Invalid ID" }); return; }
  const events = await db.select().from(timelineEventsTable)
    .where(eq(timelineEventsTable.complaintId, id))
    .orderBy(desc(timelineEventsTable.createdAt));
  res.json(events.map(e => ({ ...e, createdAt: e.createdAt.toISOString() })));
});

export default router;