# Attached-Assets/Attached-Assets/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ResolutionIQ AI - Civic Intelligence Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: complaints
    description: Complaint management
  - name: departments
    description: Department management
  - name: analytics
    description: Analytics and insights
  - name: users
    description: User management
  - name: risk
    description: Predictive risk alerts
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /complaints:
    get:
      operationId: listComplaints
      tags: [complaints]
      summary: List all complaints
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: department
          schema:
            type: string
        - in: query
          name: severity
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: ward
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of complaints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Complaint"
    post:
      operationId: createComplaint
      tags: [complaints]
      summary: Submit a new complaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintInput"
      responses:
        "201":
          description: Complaint created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"

  /complaints/{id}:
    get:
      operationId: getComplaint
      tags: [complaints]
      summary: Get complaint by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Complaint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
    patch:
      operationId: updateComplaint
      tags: [complaints]
      summary: Update complaint status or assignment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintUpdate"
      responses:
        "200":
          description: Updated complaint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"

  /complaints/{id}/vote:
    post:
      operationId: voteComplaint
      tags: [complaints]
      summary: Upvote a complaint (co-voting)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vote recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"

  /complaints/{id}/timeline:
    get:
      operationId: getComplaintTimeline
      tags: [complaints]
      summary: Get complaint timeline events
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Timeline events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineEvent"

  /complaints/stats/summary:
    get:
      operationId: getComplaintsSummary
      tags: [complaints]
      summary: Get complaint summary stats (KPIs)
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintsSummary"

  /complaints/stats/recent:
    get:
      operationId: getRecentComplaints
      tags: [complaints]
      summary: Get recent complaints for activity feed
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Recent complaints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Complaint"

  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"

  /departments/{id}:
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update department
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentUpdate"
      responses:
        "200":
          description: Updated department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"

  /departments/stats/workload:
    get:
      operationId: getDepartmentWorkload
      tags: [departments]
      summary: Get per-department complaint workload stats
      responses:
        "200":
          description: Department workload
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DepartmentWorkload"

  /analytics/overview:
    get:
      operationId: getAnalyticsOverview
      tags: [analytics]
      summary: Get platform-wide analytics overview
      responses:
        "200":
          description: Analytics overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsOverview"

  /analytics/category-breakdown:
    get:
      operationId: getCategoryBreakdown
      tags: [analytics]
      summary: Get complaint counts by category
      responses:
        "200":
          description: Category breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryCount"

  /analytics/severity-breakdown:
    get:
      operationId: getSeverityBreakdown
      tags: [analytics]
      summary: Get complaint counts by severity
      responses:
        "200":
          description: Severity breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SeverityCount"

  /analytics/ward-heatmap:
    get:
      operationId: getWardHeatmap
      tags: [analytics]
      summary: Get complaint counts per ward for heatmap
      responses:
        "200":
          description: Ward heatmap data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WardCount"

  /analytics/resolution-trends:
    get:
      operationId: getResolutionTrends
      tags: [analytics]
      summary: Get weekly resolution trend data
      responses:
        "200":
          description: Resolution trends
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendPoint"

  /analytics/sla-breach-risk:
    get:
      operationId: getSlaBreachRisk
      tags: [analytics]
      summary: Get complaints at risk of SLA breach
      responses:
        "200":
          description: SLA breach risk list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SlaRiskItem"

  /risk/alerts:
    get:
      operationId: listRiskAlerts
      tags: [risk]
      summary: List predictive infrastructure risk alerts
      responses:
        "200":
          description: Risk alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RiskAlert"
    post:
      operationId: createRiskAlert
      tags: [risk]
      summary: Create a risk alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskAlertInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskAlert"

  /risk/alerts/{id}:
    patch:
      operationId: updateRiskAlert
      tags: [risk]
      summary: Update risk alert status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskAlertUpdate"
      responses:
        "200":
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskAlert"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Complaint:
      type: object
      required: [id, title, description, category, department, status, severity, priority, ward, location, citizenName, citizenEmail, createdAt, votes, aiConfidence, aiSummary]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        department:
          type: string
        status:
          type: string
        severity:
          type: string
        priority:
          type: string
        ward:
          type: string
        location:
          type: string
        citizenName:
          type: string
        citizenEmail:
          type: string
        assignedTo:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        votes:
          type: integer
        aiConfidence:
          type: number
        aiSummary:
          type: string
        isDuplicate:
          type: boolean
        duplicateOf:
          type: ["integer", "null"]
        estimatedResolutionDays:
          type: integer
        resolvedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    ComplaintInput:
      type: object
      required: [title, description, category, ward, location, citizenName, citizenEmail]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        ward:
          type: string
        location:
          type: string
        citizenName:
          type: string
        citizenEmail:
          type: string
        imageUrl:
          type: string

    ComplaintUpdate:
      type: object
      properties:
        status:
          type: string
        assignedTo:
          type: string
        priority:
          type: string
        severity:
          type: string
        notes:
          type: string

    TimelineEvent:
      type: object
      required: [id, complaintId, eventType, description, actor, createdAt]
      properties:
        id:
          type: integer
        complaintId:
          type: integer
        eventType:
          type: string
        description:
          type: string
        actor:
          type: string
        createdAt:
          type: string

    ComplaintsSummary:
      type: object
      required: [total, pending, inProgress, resolved, avgResolutionDays, criticalCount, todayCount, satisfactionRate]
      properties:
        total:
          type: integer
        pending:
          type: integer
        inProgress:
          type: integer
        resolved:
          type: integer
        avgResolutionDays:
          type: number
        criticalCount:
          type: integer
        todayCount:
          type: integer
        satisfactionRate:
          type: number

    Department:
      type: object
      required: [id, name, code, head, email, slaHours, activeComplaints, totalResolved, avgResolutionHours]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        head:
          type: string
        email:
          type: string
        slaHours:
          type: integer
        activeComplaints:
          type: integer
        totalResolved:
          type: integer
        avgResolutionHours:
          type: number

    DepartmentInput:
      type: object
      required: [name, code, head, email, slaHours]
      properties:
        name:
          type: string
        code:
          type: string
        head:
          type: string
        email:
          type: string
        slaHours:
          type: integer

    DepartmentUpdate:
      type: object
      properties:
        name:
          type: string
        head:
          type: string
        email:
          type: string
        slaHours:
          type: integer

    DepartmentWorkload:
      type: object
      required: [departmentName, activeComplaints, resolved, pending, breachedSla]
      properties:
        departmentName:
          type: string
        activeComplaints:
          type: integer
        resolved:
          type: integer
        pending:
          type: integer
        breachedSla:
          type: integer

    AnalyticsOverview:
      type: object
      required: [totalComplaints, resolutionRate, avgResponseTimeHours, citizenSatisfaction, duplicatesBlocked, aiRoutingAccuracy, activeAlerts, wardsCovered]
      properties:
        totalComplaints:
          type: integer
        resolutionRate:
          type: number
        avgResponseTimeHours:
          type: number
        citizenSatisfaction:
          type: number
        duplicatesBlocked:
          type: integer
        aiRoutingAccuracy:
          type: number
        activeAlerts:
          type: integer
        wardsCovered:
          type: integer

    CategoryCount:
      type: object
      required: [category, count, percentage]
      properties:
        category:
          type: string
        count:
          type: integer
        percentage:
          type: number

    SeverityCount:
      type: object
      required: [severity, count]
      properties:
        severity:
          type: string
        count:
          type: integer

    WardCount:
      type: object
      required: [ward, count, severity]
      properties:
        ward:
          type: string
        count:
          type: integer
        severity:
          type: string

    TrendPoint:
      type: object
      required: [week, submitted, resolved, avgDays]
      properties:
        week:
          type: string
        submitted:
          type: integer
        resolved:
          type: integer
        avgDays:
          type: number

    SlaRiskItem:
      type: object
      required: [complaintId, title, department, hoursElapsed, slaHours, riskLevel]
      properties:
        complaintId:
          type: integer
        title:
          type: string
        department:
          type: string
        hoursElapsed:
          type: number
        slaHours:
          type: integer
        riskLevel:
          type: string

    RiskAlert:
      type: object
      required: [id, title, description, riskType, location, ward, severity, status, confidence, relatedComplaintIds, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        riskType:
          type: string
        location:
          type: string
        ward:
          type: string
        severity:
          type: string
        status:
          type: string
        confidence:
          type: number
        relatedComplaintIds:
          type: array
          items:
            type: integer
        createdAt:
          type: string
        resolvedAt:
          type: ["string", "null"]

    RiskAlertInput:
      type: object
      required: [title, description, riskType, location, ward, severity, confidence]
      properties:
        title:
          type: string
        description:
          type: string
        riskType:
          type: string
        location:
          type: string
        ward:
          type: string
        severity:
          type: string
        confidence:
          type: number
        relatedComplaintIds:
          type: array
          items:
            type: integer

    RiskAlertUpdate:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string

    User:
      type: object
      required: [id, name, email, role, department, status, complaintsHandled, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        department:
          type: string
        status:
          type: string
        complaintsHandled:
          type: integer
        createdAt:
          type: string

    UserInput:
      type: object
      required: [name, email, role, department]
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        department:
          type: string

    UserUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        department:
          type: string
        status:
          type: string