# Attached-Assets/Attached-Assets/lib/api-zod/src/generated/types/analyticsOverview.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ResolutionIQ AI - Civic Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyticsOverview {
  totalComplaints: number;
  resolutionRate: number;
  avgResponseTimeHours: number;
  citizenSatisfaction: number;
  duplicatesBlocked: number;
  aiRoutingAccuracy: number;
  activeAlerts: number;
  wardsCovered: number;
}