# Attached-Assets/Attached-Assets/lib/api-zod/src/generated/types/departmentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ResolutionIQ AI - Civic Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DepartmentInput {
  name: string;
  code: string;
  head: string;
  email: string;
  slaHours: number;
}